Isolates logic to create the database
